+++ packages/web/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.161.2](https://github.com/standardnotes/app/compare/@standardnotes/web@3.161.1...@standardnotes/web@3.161.2) (2023-06-16)
+
+**Note:** Version bump only for package @standardnotes/web
+
+## [3.161.1](https://github.com/standardnotes/app/compare/@standardnotes/web@3.161.0...@standardnotes/web@3.161.1) (2023-06-06)
+
+**Note:** Version bump only for package @standardnotes/web
+
+# [3.161.0](https://github.com/standardnotes/app/compare/@standardnotes/web@3.160.19...@standardnotes/web@3.161.0) (2023-06-02)
+
+### Features
+
+* You can now drag-n-drop notes onto tags to move them from one tag to another ([#2334](https://github.com/standardnotes/app/issues/2334)) ([08c38fc](https://github.com/standardnotes/app/commit/08c38fc7cfbce96c5de269073f5f0117000066a8))
+
+## [3.160.19](https://github.com/standardnotes/app/compare/@standardnotes/web@3.160.18...@standardnotes/web@3.160.19) (2023-05-31)
+
+### Bug Fixes
+
+* **desktop:** Fixed issue where local backup location would default to "[object Promise]" ([7d691aa](https://github.com/standardnotes/app/commit/7d691aa409af3636ad795491b8b817d9eb7a980d))
+
 ## [3.160.18](https://github.com/standardnotes/app/compare/@standardnotes/web@3.160.17...@standardnotes/web@3.160.18) (2023-05-30)
 
 ### Bug Fixes